pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cath-domain-vis"
version = "0.1.0"
description = "Lean CATH/TED domain visualization on AlphaFold structures"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "biopython>=1.83",
    "py3Dmol>=2.1.0",
    "ipywidgets>=8.0.0",
    "plotly>=5.22.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
analyses = [
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "numpy>=1.26.0",
]

[tool.setuptools.packages.find]
include = ["domain_vis*"]
